move get_entry_attribute_by_class to data store
